refactor(engine): split database.ts into focused db/ modules
Splits the 1,071-line database service monolith into 7 focused modules:
- db/types.ts (errors, type aliases)
- db/timestamps.ts (now, unixNow, generateId)
- db/query.ts (queryOne, queryMany, execute)
- db/batch.ts (batch, withSession)
- db/validation.ts (table/column name validation)
- db/crud.ts (insert, update, delete, exists, count)
- db/tenant.ts (TenantDb class, getTenantDb)

Original database.ts becomes a thin re-export barrel — all existing
import paths continue to work unchanged.
